AI Voice/dental_ai_demo/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.31.1
python-dotenv==1.0.1
jinja2==3.1.4
aiofiles==23.2.1
httpx==0.27.2
openai==1.51.2
elevenlabs==1.8.1
pydantic==2.9.2
